update user favorite publisher rules

+++ controllers/book.js
@@ -1,3 +1,4 @@
+const path = require("path")
 const asynchandler = require('express-async-handler')
 const Book = require('../models/Book')
 
@@ -31,9 +32,11 @@ const getBooks = asynchandler(async (req,res) => {
 })
 
 const setBook = asynchandler(async (req,res) => {
-    if(!req.body){
+    req.body.user = req.user.id
+    const publishedBook = await Book.findOne({user:req.user.id})
+    if (publishedBook && req.user.role !== "admin"){
         res.status(400)
-        throw new Error("Something error");
+        throw new Error(`The User Id ${req.user.id}  Has Already published a Book`) 
     }
     const book = await Book.create(req.body)
     res.status(201).json(book)
@@ -49,12 +52,18 @@ const getBook = asynchandler(async(req,res) => {
 })
 const putBook = asynchandler(async(req,res) => {
     const book = await Book.findById(req.params.id)
+    
     if(!book){
         res.status(404)
         throw new Error("Book not found") 
     }
+    if(book.user != req.user.id){
+        res.status(400)
+        throw new Error(`This not your book `)
+    }
     const updateBook = await Book.updateOne({_id:req.params.id},req.body)
     res.status(200).json(updateBook)
+    
 })
 const deleteBook = asynchandler(async(req,res) => {
     const book = await Book.findById(req.params.id)
@@ -62,13 +71,53 @@ const deleteBook = asynchandler(async(req,res) => {
         res.status(404)
         throw new Error("Book not found") 
     }
+    if(book.user != req.user.id){
+        res.status(400)
+        throw new Error(`This not your book `)
+    }
     const deletedBook = await Book.deleteOne({_id:req.params.id})
     res.status(200).json({message:'success'})
+})
+
+const uploadPhotoBook = asynchandler(async(req,res) => {
+    const book = await Book.findById(req.params.id)
+    
+    if(!book){
+        res.status(404)
+        throw new Error("Book not found") 
+    }
+    if(book.user != req.user.id){
+        res.status(400)
+        throw new Error(`This not your book `)
+    }
+    if(!req.files){
+        res.status(404)
+        throw new Error("file not upload") 
+    }
+    const file = req.files.file
+    if(!file.mimetype.startsWith('image')){
+        res.status(404)
+        throw new Error("this is not a file image") 
+    }
+    if(!file.size > process.env.MAX_FILE_UPLOAD){
+        res.status(404)
+        throw new Error(`the size of file is big than ${process.env.MAX_FILE_UPLOAD} `) 
+    }
+    file.name = `photo${book.id}${path.parse(file.name).ext}`
+    file.mv(`${process.env.FILE_UPLOAD_PATH}/${file.name}` , async error => {
+        if(error){
+            res.status(500)
+            throw new Error(` file is not upload her `) 
+        }
+        await Book.findByIdAndUpdate(req.params.id , { "volumeInfo.imageLinks.photo" : file.name,"volumeInfo.imageLinks.smallphoto" : file.name})
+        res.status(200).json({message:'success' , data : file.name})
+    })
 })
 module.exports = {
     getBooks,
     setBook,
     getBook,
     putBook,
-    deleteBook
+    deleteBook,
+    uploadPhotoBook
 }
